watchdog: ensure configured timeout is used instead of USEC_INFINITY
In some rare cases, a watchdog driver might not be able to either change
the watchdog timeout value, nor read it from within the running system.

In those situations, systemd ends up ignoring the timeout defined by the
system.conf and rather uses USEC_INFINITY. Consequently, the watchdog is
not pinged anymore and eventually resets the system.

We therefore want to ensure that the system keeps running with the
originally configured timeout.